feat(report): 비율 계산값을 정수로 반환

+++ src/main/java/hackathon/kb/chakchak/domain/report/api/dto/ReportResponseDto.java
@@ -11,8 +11,8 @@ public class ReportResponseDto {
 	private Integer totalSales;     // 총 매출액
 	private Integer successCnt;     // 성공 횟수
 	private Integer failedCnt;      // 실패 횟수
-	private Double successRate;     // 성공률 (%)
+	private Integer successRate;     // 성공률 (%)
 
-	private Map<String, Double> ageDistribution;   // 연령대 분포 (%)
-	private Map<String, Double> genderDistribution; // 성별 분포 (%)
+	private Map<String, Integer> ageDistribution;   // 연령대 분포 (%)
+	private Map<String, Integer> genderDistribution; // 성별 분포 (%)
 }
